Move plugin to its own module
